+++ test/Sentry.AspNetCore.Tests/ModuleInit.cs
@@ -1,10 +1,12 @@
+#if NET6_0
+
 using System.Runtime.CompilerServices;
 
 public static class ModuleInit
 {
     [ModuleInitializer]
-    public static void Init()
-    {
-        VerifyDiffPlex.Initialize();
-    }
+    public static void Init() =>
+        VerifyHttp.Enable();
 }
+
+#endif